Add analysis package and cleanup util package
- move results.py to new file plot_stats.py and make compatible with model object code
- move analysis.py to new file stats.py
- add subtitle to plots
- make sure space size param l is float everywhere

+++ flock/vicsek.py
@@ -35,7 +35,7 @@ class VicsekModel(FlockModel):
     """
 
     def __init__(self,
-                 n: int, l: int,
+                 n: int, l: float,
                  bounds: EnumBounds, neighbours: EnumNeighbours,
                  e: float, v: float = 0.3, r: float = 1, dt: float = 1) -> None:
         """
